+++ src/modules/rpc/server_ban.c
@@ -248,7 +248,7 @@ RPC_CALL_FUNC(rpc_server_ban_del)
 	tkllayer[4] = hostmask;
 	tkllayer[5] = set_by;
 	tkllayer[6] = NULL;
-	cmd_tkl(&me, NULL, 6, tkllayer);
+	cmd_tkl(NULL, &me, NULL, 6, tkllayer);
 
 	if (!find_tkl_serverban(tkl_type_int, usermask, hostmask, soft))
 	{